refactor(useField): utils.ts -> utilities.ts
